change loggedAt type from integer to bigInterger

+++ migrations/20230827170315_add-logged-at-column.js
@@ -1,5 +1,5 @@
 exports.up = (knex) => knex.schema.alterTable("User", (table) => {
-  table.integer("loggedAt");
+  table.bigInteger("loggedAt");
 });
 
 exports.down = (knex) => knex.schema.alterTable("User", (table) => {
